refactor(database): add `DatabaseManager` and `page.db` entity
