+++ packages/blocks/src/database-block/components/edit-column-popup.ts
@@ -347,10 +347,10 @@ export class EditColumnPopup extends LitElement {
     columnId: string,
     schemaProperties: Partial<ColumnSchema>
   ) => {
-    const currentSchema = this.targetModel.page.getColumnSchema(columnId);
+    const currentSchema = this.targetModel.page.db.getColumnSchema(columnId);
     assertExists(currentSchema);
     const schema = { ...currentSchema, ...schemaProperties };
-    this.targetModel.page.updateColumnSchema(schema);
+    this.targetModel.page.db.updateColumnSchema(schema);
   };
 
   private _changeColumnType = (
@@ -369,12 +369,12 @@ export class EditColumnPopup extends LitElement {
     // multi-select -> select
     else if (currentType === 'multi-select' && targetType === 'select') {
       this._updateColumnSchema(columnId, { type: targetType });
-      this.targetModel.page.convertColumn(columnId, 'select');
+      this.targetModel.page.db.convertColumn(columnId, 'select');
     }
     // number -> rich-text
     else if (currentType === 'number' && targetType === 'rich-text') {
       this._updateColumnSchema(columnId, { type: targetType });
-      this.targetModel.page.convertColumn(columnId, 'rich-text');
+      this.targetModel.page.db.convertColumn(columnId, 'rich-text');
     } else {
       // incompatible types: clear the value of the column
       const renderer = getColumnSchemaRenderer(targetType);
@@ -382,7 +382,7 @@ export class EditColumnPopup extends LitElement {
         type: targetType,
         property: renderer.propertyCreator(),
       });
-      this.targetModel.page.deleteColumn(columnId);
+      this.targetModel.page.db.deleteColumn(columnId);
     }
 
     this.closePopup();
@@ -406,8 +406,8 @@ export class EditColumnPopup extends LitElement {
 
     if (actionType === 'delete') {
       this.targetModel.page.captureSync();
-      this.targetModel.page.deleteColumnSchema(columnId);
-      this.targetModel.page.deleteColumn(columnId);
+      this.targetModel.page.db.deleteColumnSchema(columnId);
+      this.targetModel.page.db.deleteColumn(columnId);
       const columns = this.targetModel.columns.filter(id => id !== columnId);
       this.targetModel.page.updateBlock(this.targetModel, {
         columns,
@@ -436,17 +436,17 @@ export class EditColumnPopup extends LitElement {
 
     if (actionType === 'duplicate') {
       this.targetModel.page.captureSync();
-      const currentSchema = this.targetModel.page.getColumnSchema(columnId);
+      const currentSchema = this.targetModel.page.db.getColumnSchema(columnId);
       assertExists(currentSchema);
       const { id: copyId, ...nonIdProps } = currentSchema;
       const schema = { ...nonIdProps };
-      const id = this.targetModel.page.updateColumnSchema(schema);
+      const id = this.targetModel.page.db.updateColumnSchema(schema);
       const newColumns = [...this.targetModel.columns];
       newColumns.splice(this.columnIndex + 1, 0, id);
       this.targetModel.page.updateBlock(this.targetModel, {
         columns: newColumns,
       });
-      this.targetModel.page.copyColumn(copyId, id);
+      this.targetModel.page.db.copyColumn(copyId, id);
       this.closePopup();
       return;
     }
