+++ packages/playground/src/data/index.ts
@@ -136,21 +136,21 @@ export const database: InitFn = (workspace: Workspace) => {
     { value: 'TODO', color: '#FFE1E1' },
     { value: 'WIP', color: '#E1EFFF' },
   ];
-  const col1 = page.updateColumnSchema({
+  const col1 = page.db.updateColumnSchema({
     name: 'Number',
     type: 'number',
     width: 200,
     hide: false,
     decimal: 0,
   });
-  const col2 = page.updateColumnSchema({
+  const col2 = page.db.updateColumnSchema({
     name: 'Single Select',
     type: 'select',
     width: 200,
     hide: false,
     selection,
   });
-  const col3 = page.updateColumnSchema({
+  const col3 = page.db.updateColumnSchema({
     name: 'Rich Text',
     type: 'rich-text',
     width: 200,
@@ -180,12 +180,12 @@ export const database: InitFn = (workspace: Workspace) => {
     databaseId
   );
 
-  page.updateColumn(p1, {
+  page.db.updateColumn(p1, {
     columnId: col1,
     value: 0.1,
   });
 
-  page.updateColumn(p2, {
+  page.db.updateColumn(p2, {
     columnId: col2,
     value: [selection[1]],
   });
@@ -193,7 +193,7 @@ export const database: InitFn = (workspace: Workspace) => {
   const text = new page.YText();
   text.insert(0, '123');
   text.insert(0, 'code');
-  page.updateColumn(p2, {
+  page.db.updateColumn(p2, {
     columnId: col3,
     value: text,
   });
